# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "description": "새로운 사용자 계정을 생성합니다.",
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "회원가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "회원가입이 완료되었습니다."
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "409": {
            "description": "이미 존재하는 이메일"
          }
        },
        "summary": "회원가입",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "사용자 인증을 통해 JWT 토큰을 발급받습니다.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "로그인이 완료되었습니다."
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청 데이터"
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "summary": "로그인",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/select-profile/{profileId}": {
      "post": {
        "description": "사용자가 특정 프로필을 선택하여 해당 프로필의 토큰을 발급받습니다.",
        "operationId": "AuthController_selectProfile",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "description": "선택할 프로필 ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "프로필 선택 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "프로필이 선택되었습니다."
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          },
          "403": {
            "description": "프로필 접근 권한 없음"
          },
          "404": {
            "description": "프로필을 찾을 수 없음"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "프로필 선택",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Thinking API Server",
    "description": "수업 관리 시스템 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "JWT 토큰을 입력하세요",
        "in": "header"
      }
    },
    "schemas": {
      "SignupPayload": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "사용자 이메일 주소",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "사용자 비밀번호 (최소 6자)",
            "example": "password123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginPayload": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "사용자 이메일 주소",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "사용자 비밀번호",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    }
  }
}